% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEMSS.R
\name{compute_kernel}
\alias{compute_kernel}
\title{Correlation function of selected type, supporting both isotropic and product forms}
\usage{
compute_kernel(x1, x2, theta, type)
}
\arguments{
\item{theta}{p by 1 vector of lengthscale parameters}

\item{type}{one of "\code{Gaussian}", "\code{Matern5_2}", "\code{Matern3_2}"}

\item{X1}{n by p matrix of input locations, one observation per row.}

\item{X2}{matrix of design locations if correlation is calculated between \code{X1} and \code{X2} (otherwise calculated between \code{X1} and itself)}
}
\description{
Correlation function of selected type, supporting both isotropic and product forms
}
\details{
Definition of univariate correlation function and hyperparameters:
\itemize{
\item "\code{Gaussian}": \eqn{c(x, y) = exp(-(x-y)^2/theta)}
\item "\code{Matern5_2}": \eqn{c(x, y) = (1+sqrt(5)/theta * abs(x-y) + 5/(3*theta^2)(x-y)^2) * exp(-sqrt(5)*abs(x-y)/theta)}
\item "\code{Matern3_2}": \eqn{c(x, y) = (1+sqrt(3)/theta * abs(x-y)) * exp(-sqrt(3)*abs(x-y)/theta)}
}
Multivariate correlations are product of univariate ones.
}
