% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEMSS.R
\name{gp_predict}
\alias{gp_predict}
\title{Gaussian Process Prediction}
\usage{
gp_predict(X_new, X, Y, Cov_fun, Parameters)
}
\arguments{
\item{X_new}{An \code{m x p} matrix of new design locations to predict.}

\item{X}{An \code{n x p} matrix of training design locations.}

\item{Y}{An \code{n x 1} vector of observed responses at training locations.}

\item{Cov_fun}{Covariance kernel type, either "Gaussian", "Matern5_2", or "Matern3_2".}

\item{Parameters}{A list of hyperparameters containing \code{theta}, \code{g}, \code{sigma2}, and \code{beta0}}
}
\value{
A list containing:
\itemize{
\item \code{mean}: A vector of predicted means at \code{X_new}.
\item \code{sd2}: A vector of predicted variances (posterior variances) at \code{X_new}.
}
}
\description{
Performs Gaussian Process prediction for a given set of new design locations,
using hyperparameters provided in the Parameters list.
}
\details{
The function uses the standard GP prediction formulas:
\deqn{\mu(x_{new}) = \beta_0 + k(x_{new}, X) K^{-1} (Y - \beta_0)}
\deqn{s^2(x_{new}) = \sigma^2 [ (1 + g) - k(x_{new}, X) (K + gI)^{-1} k(X, x_{new}) ]}
}
